fix `brainpy.math.softplus` and `brainpy.dnn.SoftPlus` (#581)
* add `normalize` parameter in dual exponential model

* fix `brainpy.math.softplus` and `brainpy.dnn.SoftPlus`

* increase default threshold to 40 in `brainpy.math.softplus`

* update the `brainpy.math.softplus`

* update requirements

* update

+++ brainpy/_src/dnn/activations.py
@@ -840,10 +840,10 @@ class Softplus(Layer):
       >>> output = m(input)
   """
   __constants__ = ['beta', 'threshold']
-  beta: int
-  threshold: int
+  beta: float
+  threshold: float
 
-  def __init__(self, beta: int = 1, threshold: int = 20) -> None:
+  def __init__(self, beta: float = 1, threshold: float = 20.) -> None:
     super().__init__()
     self.beta = beta
     self.threshold = threshold
